+++ lib/clixon/clixon_xml_changelog.h
@@ -40,8 +40,8 @@
 /*
  * Prototypes
  */
-int xml_changelog_upgrade(clicon_handle h, cxobj *xn, char *namespace, uint32_t from, uint32_t to, void *arg, cbuf *cbret);
+int xml_changelog_upgrade(clicon_handle h, cxobj *xn, char *ns, uint16_t op, uint32_t from, uint32_t to, void *arg, cbuf *cbret);
 int clixon_xml_changelog_init(clicon_handle h);
-int xml_namespace_vec(clicon_handle h, cxobj *xt, char *namespace, cxobj ***vec, size_t *veclen);
+int xml_namespace_vec(clicon_handle h, cxobj *xt, char *ns, cxobj ***vec, size_t *veclen);
 
 #endif /* _CLIXON_XML_CHANGELOG_H */