+++ lib/clixon/clixon_yang_module.h
@@ -54,8 +54,7 @@
 typedef struct {
     int    md_status; /* 0 if no module-state in a datastore, 1 if there is */
     char  *md_set_id; /* server-specific identifier */
-    cxobj *md_del;    /* yang module state deletes */
-    cxobj *md_mod;    /* yang module state modifications */
+    cxobj *md_diff;  /* yang module state containing revisions and XML_FLAG_ADD|DEL|CHANGE */
 } modstate_diff_t;
 
 /*
@@ -72,7 +71,7 @@ int yang_modules_state_get(clicon_handle h, yang_stmt *yspec, char *xpath,
 int clixon_module_upgrade(clicon_handle h, cxobj *xt, modstate_diff_t *msd, cbuf *cb);
 yang_stmt *yang_find_module_by_prefix(yang_stmt *ys, char *prefix);
 yang_stmt *yang_find_module_by_prefix_yspec(yang_stmt *yspec, char *prefix);
-yang_stmt *yang_find_module_by_namespace(yang_stmt *yspec, char *namespace);
+yang_stmt *yang_find_module_by_namespace(yang_stmt *yspec, const char *ns);
 yang_stmt *yang_find_module_by_name(yang_stmt *yspec, char *name);
 
 #endif  /* _CLIXON_YANG_MODULE_H_ */