+++ lib/src/clixon_xml_changelog.c
@@ -355,7 +355,8 @@ changelog_iterate(clicon_handle h,
 /*! Automatic upgrade using changelog
  * @param[in]  h       Clicon handle 
  * @param[in]  xt      Top-level XML tree to be updated (includes other ns as well)
- * @param[in]  namespace Namespace of module (for info)
+ * @param[in]  ns      Namespace of module (for info)
+ * @param[in]  op      One of XML_FLAG_ADD, _DEL, _CHANGE
  * @param[in]  from    From revision on the form YYYYMMDD
  * @param[in]  to      To revision on the form YYYYMMDD (0 not in system)
  * @param[in]  arg     User argument given at rpc_callback_register() 
@@ -367,12 +368,13 @@ changelog_iterate(clicon_handle h,
  */
 int
 xml_changelog_upgrade(clicon_handle h,       
-		       cxobj        *xt,      
-		       char         *namespace,
-		       uint32_t      from,
-		       uint32_t      to,
-		       void         *arg,     
-		       cbuf         *cbret)
+		      cxobj        *xt,      
+		      char         *ns,
+		      uint16_t      op,
+		      uint32_t      from,
+		      uint32_t      to,
+		      void         *arg,     
+		      cbuf         *cbret)
 {
     int        retval = -1;
     cxobj     *xchlog; /* changelog */
@@ -397,7 +399,7 @@ xml_changelog_upgrade(clicon_handle h,
      * - note it t=0 then no changelog is applied
      */
     if (xpath_vec(xchlog, NULL, "changelog[namespace=\"%s\"]", 
-		  &vec, &veclen, namespace) < 0)
+		  &vec, &veclen, ns) < 0)
 	goto done;
     /* Get all changelogs in the interval [from,to]*/
     for (i=0; i<veclen; i++){
@@ -485,7 +487,7 @@ clixon_xml_changelog_init(clicon_handle h)
 /*! Given a top-level XML tree and a namespace, return a vector of matching XML nodes
  * @param[in]  h         Clicon handle
  * @param[in]  xt        Top-level XML tree, with children marked with namespaces
- * @param[in]  namespace The namespace to select
+ * @param[in]  ns		 The namespace to select
  * @param[out] vecp      Vector containining XML nodes w namespace. Null-terminated.
  * @param[out] veclenp   Length of vector
  * @note  Need to free vec after use with free()
@@ -498,7 +500,7 @@ clixon_xml_changelog_init(clicon_handle h)
 int
 xml_namespace_vec(clicon_handle h,
 		  cxobj        *xt,
-		  char         *namespace,
+		  char         *ns,
 		  cxobj      ***vecp,
 		  size_t       *veclenp)
 {
@@ -506,7 +508,7 @@ xml_namespace_vec(clicon_handle h,
     cxobj **xvec = NULL;
     size_t  xlen;
     cxobj  *xc;
-    char   *ns;
+    char   *ns0;
     int     i;
 
     /* Allocate upper bound on length (ie could be too large) + a NULL element
@@ -521,9 +523,9 @@ xml_namespace_vec(clicon_handle h,
     xc = NULL;
     i = 0;
     while ((xc = xml_child_each(xt, xc, CX_ELMNT)) != NULL) {
-	if (xml2ns(xc, NULL, &ns) < 0) /* Get namespace of XML */
+	if (xml2ns(xc, NULL, &ns0) < 0) /* Get namespace of XML */
 	    goto done;       
-	if (strcmp(namespace, ns))
+	if (strcmp(ns, ns0))
 	    continue; /* no match */
 	xvec[i++] = xc;
     }